refactor: ajust layout home
