+++ src/app/home/layout.jsx
@@ -1,0 +1,5 @@
+const LayoutRoot = ({ children }) => {
+  return <div className="w-full h-full">{children}</div>;
+};
+
+export default LayoutRoot;
