+++ src/app/layout.js
@@ -1,7 +1,6 @@
 import { Archivo } from 'next/font/google';
 import { Providers } from '@/contexts/Providers';
 import './globals.css';
-// import 'rsuite/dist/rsuite.min.css';
 
 const inter = Archivo({ subsets: ['latin'] });
